+++ CHANGELOG.md
@@ -6,7 +6,16 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.2.0 - [Unreleased]
+## 0.2.1 - 2021-04-17
+
+### Added
+
+- Wrapped new ACL-related functions in `Spear.Client`
+    - `c:Spear.Client.get_stream_metadata/2`
+    - `c:Spear.Client.set_stream_metadata/2`
+    - `c:Spear.Client.set_global_acl/3`
+
+## 0.2.0 - 2021-04-17
 
 ### Changed